voicesBack/requirements.txt
Django==2.1.5
django-cors-headers==2.4.0
django-cuser==2017.3.16
django-rest-auth==0.9.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
Pillow==5.4.1
psycopg2==2.7.7
PyJWT==1.7.1
python-crontab==2.3.6
python-dateutil==2.8.0
python-http-client==3.1.0
pytz==2018.9
sendgrid==5.6.0
six==1.12.0
